requirements.txt
asgiref==3.5.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-serializer-extensions==2.0.1
hashids==1.3.1
inflection==0.5.1
itsdangerous==2.1.2
Pillow==9.4.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
